@@ -2,6 +2,7 @@
   "name": "svelte",
   "version": "4.0.0-next.0",
   "description": "Cybernetically enhanced web apps",
+  "type": "module",
   "module": "index.mjs",
   "main": "index",
   "files": [
@@ -89,7 +90,7 @@
     "dev": "rollup -cw",
     "posttest": "agadoo internal/index.mjs",
     "prepublishOnly": "node check_publish_env.js && npm run lint && npm run build && npm test",
-    "tsd": "tsc -p src/compiler --emitDeclarationOnly && tsc -p src/runtime --emitDeclarationOnly",
+    "tsd": "node ./generate-types.mjs",
     "lint": "eslint \"{src,test}/**/*.{ts,js}\" --cache"
   },
   "repository": {
@@ -119,23 +120,24 @@
     "@rollup/plugin-sucrase": "^5.0.1",
     "@rollup/plugin-typescript": "^11.1.0",
     "@rollup/plugin-virtual": "^3.0.1",
-    "@sveltejs/eslint-config": "github:sveltejs/eslint-config#v5.8.0",
+    "@sveltejs/eslint-config": "^6.0.1",
     "@types/aria-query": "^5.0.1",
     "@types/estree": "^1.0.0",
     "@types/mocha": "^10.0.1",
     "@types/node": "^14.14.31",
     "@typescript-eslint/eslint-plugin": "^5.58.0",
-    "@typescript-eslint/parser": "^5.58.0",
     "acorn": "^8.8.2",
     "agadoo": "^3.0.0",
     "aria-query": "^5.1.3",
     "axobject-query": "^3.1.1",
     "code-red": "^1.0.0",
     "css-tree": "^2.3.1",
-    "eslint": "^8.35.0",
+    "esbuild": "^0.17.19",
+    "eslint": "^8.40.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-svelte3": "^4.0.0",
+    "eslint-plugin-svelte": "^2.28.0",
+    "eslint-plugin-unicorn": "^47.0.0",
     "estree-walker": "^3.0.3",
     "happy-dom": "^9.18.3",
     "is-reference": "^3.0.1",
@@ -147,13 +149,14 @@
     "prettier": "^2.8.8",
     "prettier-plugin-svelte": "^2.10.0",
     "rollup": "^3.20.2",
+    "rollup-plugin-dts": "^5.3.0",
     "source-map": "^0.7.4",
     "source-map-support": "^0.5.21",
     "tiny-glob": "^0.2.9",
     "tslib": "^2.5.0",
     "typescript": "^5.0.4",
     "util": "^0.12.5",
-    "vitest": "^0.31.0"
+    "vitest": "^0.31.1"
   },
   "packageManager": "pnpm@8.4.0"
 }
