@@ -1,8 +1,8 @@
 // this file will replace all the output.json files with their _actual.json
 // equivalents. Only use it when you're sure that you haven't
 // broken anything!
-const fs = require('fs');
-const glob = require('tiny-glob/sync.js');
+import * as fs from 'node:fs';
+import glob from 'tiny-glob/sync.js';
 
 glob('samples/*/_actual.json', { cwd: __dirname }).forEach((file) => {
 	const actual = fs.readFileSync(`${__dirname}/${file}`, 'utf-8');
