@@ -32,7 +32,7 @@ const runtime_entrypoints = Object.fromEntries(
 	fs
 		.readdirSync('src/runtime', { withFileTypes: true })
 		.filter((dirent) => dirent.isDirectory())
-		.map((dirent) => [dirent.name, `src/runtime/${dirent.name}/index.ts`])
+		.map((dirent) => [dirent.name, `src/runtime/${dirent.name}/index.js`])
 );
 
 /**
@@ -42,8 +42,8 @@ export default [
 	{
 		input: {
 			...runtime_entrypoints,
-			index: 'src/runtime/index.ts',
-			ssr: 'src/runtime/ssr.ts'
+			index: 'src/runtime/index.js',
+			ssr: 'src/runtime/ssr.js'
 		},
 		output: ['es', 'cjs'].map(
 			/** @returns {import('rollup').OutputOptions} */
@@ -85,7 +85,7 @@ export default [
 							const mod = bundle[`internal/index.mjs`];
 							if (mod) {
 								fs.writeFileSync(
-									'src/compiler/compile/internal_exports.ts',
+									'src/compiler/compile/internal_exports.js',
 									`// This file is automatically generated\n` +
 										`export default new Set(${JSON.stringify(mod.exports)});`
 								);
@@ -103,7 +103,7 @@ export default [
 	},
 	/* compiler.js */
 	{
-		input: 'src/compiler/index.ts',
+		input: 'src/compiler/index.js',
 		plugins: [
 			replace({
 				preventAssignment: true,
