@@ -1,8 +1,8 @@
-import * as fs from 'fs';
-import * as path from 'path';
+import * as fs from 'node:fs';
+import * as path from 'node:path';
 import { describe, it, assert } from 'vitest';
-import { try_load_config, should_update_expected } from '../helpers';
-import * as svelte from '../../compiler';
+import { try_load_config, should_update_expected } from '../helpers.js';
+import * as svelte from 'svelte/compiler';
 
 describe('js-output', () => {
 	fs.readdirSync(`${__dirname}/samples`).forEach((dir) => {
@@ -33,14 +33,11 @@ describe('js-output', () => {
 			let actual;
 
 			try {
-				const options = Object.assign(config.options || {});
+				const options = Object.assign({}, config.options || {});
 
 				actual = svelte
 					.compile(input, options)
-					.js.code.replace(
-						/generated by Svelte v\d+\.\d+\.\d+(-\w+\.\d+)?/,
-						'generated by Svelte vX.Y.Z'
-					);
+					.js.code.replace(/generated by Svelte v__VERSION__/, 'generated by Svelte vX.Y.Z');
 			} catch (err) {
 				console.log(err.frame);
 				throw err;
