@@ -1,7 +1,7 @@
-import * as fs from 'fs';
+import * as fs from 'node:fs';
 import { assert, describe, it } from 'vitest';
-import * as svelte from '../../compiler';
-import { try_load_json } from '../helpers';
+import * as svelte from 'svelte/compiler';
+import { try_load_json } from '../helpers.js';
 
 describe('parse', () => {
 	fs.readdirSync(`${__dirname}/samples`).forEach((dir) => {
@@ -32,7 +32,7 @@ describe('parse', () => {
 			try {
 				const { ast } = svelte.compile(
 					input,
-					Object.assign(options, {
+					Object.assign({}, options, {
 						generate: false
 					})
 				);
