@@ -4,8 +4,8 @@
 // TODO: Isolate the html comparison
 // TODO: happy-dom might be faster but currently replaces quotes which fails assertions
 
-import * as fs from 'fs';
-import * as path from 'path';
+import * as fs from 'node:fs';
+import * as path from 'node:path';
 import { assert, describe, it } from 'vitest';
 import {
 	create_loader,
@@ -14,7 +14,7 @@ import {
 	try_load_json,
 	try_read_file
 } from '../helpers.js';
-import { assert_html_equal } from '../html_equal.js';
+import { assert_html_equal_with_options } from '../html_equal.js';
 
 describe('ssr', async () => {
 	async function run_test(dir) {
@@ -50,9 +50,9 @@ describe('ssr', async () => {
 			if (css.code) fs.writeFileSync(`${dir}/_actual.css`, css.code);
 
 			try {
-				assert_html_equal(html, expectedHtml, {
-					normalize_html: { preserveComments: compileOptions.preserveComments },
-					without_normalize: config.withoutNormalizeHtml
+				assert_html_equal_with_options(html, expectedHtml, {
+					preserveComments: compileOptions.preserveComments,
+					withoutNormalizeHtml: config.withoutNormalizeHtml
 				});
 			} catch (error) {
 				if (should_update_expected()) {
@@ -83,9 +83,13 @@ describe('ssr', async () => {
 				fs.writeFileSync(`${dir}/_actual-head.html`, head);
 
 				try {
-					assert_html_equal(head, fs.readFileSync(`${dir}/_expected-head.html`, 'utf-8'), {
-						normalize_html: { preserveComments: compileOptions.hydratable }
-					});
+					assert_html_equal_with_options(
+						head,
+						fs.readFileSync(`${dir}/_expected-head.html`, 'utf-8'),
+						{
+							preserveComments: compileOptions.hydratable
+						}
+					);
 				} catch (error) {
 					if (should_update_expected()) {
 						fs.writeFileSync(`${dir}/_expected-head.html`, head);
