@@ -1,7 +1,7 @@
-import * as fs from 'fs';
+import * as fs from 'node:fs';
 import { describe, it, assert } from 'vitest';
-import * as svelte from '../../compiler';
-import { try_load_config, try_load_json } from '../helpers';
+import * as svelte from 'svelte/compiler';
+import { try_load_config, try_load_json } from '../helpers.js';
 
 describe('stats', () => {
 	fs.readdirSync(`${__dirname}/samples`).forEach((dir) => {
