@@ -1,11 +1,13 @@
-import * as fs from 'fs';
+import * as fs from 'node:fs';
 import { assert, describe, it } from 'vitest';
-import { compile } from '../../compiler.mjs';
+import { compile } from 'svelte/compiler';
 import { try_load_json } from '../helpers.js';
 
 describe('vars', () => {
 	fs.readdirSync(`${__dirname}/samples`).forEach((dir) => {
-		if (dir[0] === '.') return;
+		if (dir[0] === '.') {
+			return;
+		}
 
 		// add .solo to a sample directory name to only run that test
 		const solo = /\.solo/.test(dir);
